Fixes to lotto models and model test cases
The code that was previously here was mostly boilerplate and ignored
using the base class. The model classes are now properly inheriting from
the base class which has also been switched from being an old style
class to a new style class.

The switch of the base class from old style class to new style class has
been done to allow the use of the super functions. Classes that do not
have tests written for them have been removed for the meantime.

The model tests now include a sample of the JSON that is used in the
test setup. The base test has been also removed for now.

TL;DR : A whole bunch of house cleaning regarding the models.